pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "stratagent"
version = "0.1.0"
description = "A Multi-Agent RAG System for Strategic Business Analysis"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "streamlit>=1.40.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pdfplumber>=0.11.0",
    "pandas>=2.0.0",
    "beautifulsoup4>=4.12.0",
    "pinecone>=5.0.0",
    "tiktoken>=0.7.0",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "pypdf>=6.7.1",
    "pytest>=9.0.2",
    "reportlab>=4.4.10",
    "langchain-huggingface>=1.2.0",
    "langchain-pinecone>=0.2.13",
    "sentence-transformers>=5.2.3",
    "yfinance>=1.2.0",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "flagembedding>=1.3.5",
    "huggingface-hub[hf-xet]>=0.36.2",
    "crewai>=1.6.1",
    "crewai-tools>=1.6.1",
    "tavily-python>=0.7.21",
    "langchain-groq>=1.1.2",
    "litellm>=1.0.0",
    "groq>=0.37.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.8.0",
    "black>=24.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
    "reportlab>=4.0.0",
]

[project.scripts]
stratagent-api = "api.main:main"
stratagent-ingest = "scripts.ingest:main"

[tool.hatch.build.targets.wheel]
packages = ["agents", "config", "ingestion", "retrieval", "api", "evaluation"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
